requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==23.3.0
click==8.1.3
decorator==5.1.1
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.4.1
environ==1.0
executing==1.2.0
flake8==6.0.0
importlib-resources==5.12.0
inflection==0.5.1
ipython==8.12.0
jedi==0.18.2
jsonschema==4.17.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
phonenumbers==8.13.8
pickleshare==0.7.5
pkgutil-resolve-name==1.3.10
platformdirs==3.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing-extensions==4.5.0
uritemplate==4.1.1
wcwidth==0.2.6
zipp==3.15.0
